webapp/requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
wsgiref==0.1.2
redis==2.10.1
python-etcd==0.3.2
